# api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/registrations:
    get:
      description: >-
        Con `?marathonId=` sirve para retomar un borrador cuando la app perdio el id. No recotiza:
        devuelve los totales guardados.
      operationId: RegistrationsController_list
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - draft
              - pending_payment
              - confirmed
              - cancelled
              - refunded
        - name: marathonId
          required: false
          in: query
          description: Id o slug de la maraton, para retomar un borrador
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationDto'
      security:
        - access-token: []
      summary: Mis inscripciones
      tags:
        - registrations
    post:
      description: >-
        Si ya hay un borrador para esa maraton lo devuelve actualizado en vez de crear otro. Si ya
        hay una inscripcion confirmada responde ALREADY_REGISTERED.
      operationId: RegistrationsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationDto'
        '409':
          description: ALREADY_REGISTERED / MARATHON_FULL / REGISTRATION_CLOSED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - access-token: []
      summary: 'Paso 1: crea el borrador con los datos personales'
      tags:
        - registrations
...
  /api/v1/marathons:
    get:
      description: >-
        Paginado por cursor y ordenado por fecha de largada. Con `status`, una pagina puede venir
        corta o vacia y traer `meta.nextCursor` igual: hay que paginar hasta que sea null.
      operationId: MarathonsController_list
      parameters:
        - name: city
          required: false
          in: query
          description: Coincidencia parcial, sin acentos ni mayusculas
          schema:
            example: La Paz
            type: string
        - name: from
          required: false
          in: query
          description: Largada desde (ISO). Inclusivo
          schema:
            example: '2026-09-01'
            type: string
        - name: to
          required: false
          in: query
          description: Largada hasta (ISO). Inclusivo
          schema:
            example: '2026-12-31'
            type: string
        - name: status
          required: false
          in: query
          description: >-
            Filtra por el estado YA RESUELTO. Como se deriva al leer, el filtrado se hace despues de
            traer las filas: por eso no se combina con paginas enormes.
          schema:
            type: string
            enum:
              - open
              - closing_soon
              - full
              - closed
        - name: limit
          required: false
          in: query
          schema:
            maximum: 50
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          description: El `meta.nextCursor` de la pagina anterior. Opaco.
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarathonSummaryDto'
      summary: Catalogo de maratones publicadas
      tags:
        - marathons
...
  /api/v1/marathons/{slug}:
    get:
      description: >-
        Acepta el slug o el id. Trae categorias y extras embebidos: son lo que pinta la pantalla de
        detalle, y pedirlos aparte serian tres viajes para una sola vista.
      operationId: MarathonsController_detail
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            example: maraton-la-paz-3600
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarathonDetailDto'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      summary: Detalle de una maraton
      tags:
        - marathons
  /api/v1/marathons/{id}/categories:
    get:
      description: Acepta el slug o el id. Existe suelto para el paso 2 de la inscripcion.
      operationId: MarathonsController_categories
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarathonCategoryDto'
      summary: Categorias de una maraton
      tags:
        - marathons
  /api/v1/marathons/{id}/extras:
    get:
      description: 'Acepta el slug o el id. `stock: null` significa sin limite, no agotado.'
      operationId: MarathonsController_extras
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarathonExtraDto'
      summary: Extras comprables de una maraton
      tags:
        - marathons
...
  /api/v1/routes:
    get:
      description: 'Sin geometria: es la lista con la que se elige uno. Ordenados por ciudad y distancia.'
      operationId: RoutesController_list
      parameters:
        - name: city
          required: false
          in: query
          description: Coincidencia parcial
          schema:
            example: La Paz
            type: string
        - name: includeArchived
          required: false
          in: query
          description: >-
            Incluye los archivados. Solo sirve para el panel: la app nunca ofrece un recorrido
            retirado para una carrera nueva.
          schema:
            default: false
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteSummaryDto'
      summary: Recorridos disponibles
      tags:
        - routes
  /api/v1/routes/{id}:
    get:
      description: >-
        Acepta el slug o el id. El `LineString` viene simplificado para dibujarlo en el movil; con
        `?full=true` sale tal como se cargo.
      operationId: RoutesController_detail
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: circuito-zona-sur
            type: string
        - name: full
          required: false
          in: query
          description: Devuelve la geometria tal como se cargo, sin simplificar. Para el panel.
          schema:
            default: false
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetailDto'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      summary: Un recorrido con su trazado
      tags:
        - routes
...
  /api/v1/admin/service-fee:
    get:
      operationId: AdminController_verFees
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Config global del cargo por servicio y las maratones que la sobrescriben
      tags:
        - admin
    put:
      description: >-
        Con `enabled: false` el total deja de llevar cargo **y la línea desaparece** de la respuesta
        de cotización: un "Bs 0,00" promete un cargo que hoy no se cobra.
      operationId: AdminController_guardarFeeGlobal
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceFeeConfigDto'
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Activar, desactivar o reconfigurar el cargo global
      tags:
        - admin
  /api/v1/admin/service-fee/preview:
    get:
      description: >-
        Se calcula con la **misma** función que cobra de verdad, así que no puede desviarse. Sin
        `subtotalCents` usa Bs 200 de ejemplo.
      operationId: AdminController_previsualizar
      parameters:
        - name: subtotalCents
          required: false
          in: query
          description: Subtotal de ejemplo, en centavos
          schema:
            example: 20000
            type: number
        - name: marathonId
          required: false
          in: query
          description: Sin esto se previsualiza la config global
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeePreviewDto'
      security:
        - access-token: []
      summary: Vista previa del efecto sobre un total de ejemplo
      tags:
        - admin
  /api/v1/admin/marathons/{id}/service-fee:
    put:
      description: >-
        El override manda **aunque venga apagado**: una maratón con una config `enabled: false` no
        cobra cargo, en vez de caer de vuelta a la global.
      operationId: AdminController_guardarFeeDeMaraton
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceFeeConfigDto'
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Override del cargo para una maratón
      tags:
        - admin
    delete:
      operationId: AdminController_quitarFeeDeMaraton
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Quitar el override y volver a la config global
      tags:
        - admin
  /api/v1/admin/routes:
    get:
      description: >-
        Es lo que llena el selector de recorrido del formulario de alta. Con `includeArchived=false`
        salen solo los que se pueden usar en una carrera nueva.
      operationId: AdminController_listarRecorridos
      parameters:
        - name: city
          required: false
          in: query
          description: Coincidencia parcial
          schema:
            example: La Paz
            type: string
        - name: includeArchived
          required: false
          in: query
          description: >-
            Incluye los archivados. Solo sirve para el panel: la app nunca ofrece un recorrido
            retirado para una carrera nueva.
          schema:
            default: false
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteSummaryDto'
      security:
        - access-token: []
      summary: Recorridos cargados, archivados incluidos
      tags:
        - admin
    post:
      description: >-
        La distancia **no se manda**: se mide sobre la geometría. Es la que después hereda cada
        maratón que lo elija.
      operationId: AdminController_crearRecorrido
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteSummaryDto'
        '400':
          description: 'VALIDATION_ERROR: el GeoJSON no es un LineString utilizable'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - access-token: []
      summary: Cargar un recorrido
      tags:
        - admin
...
  /api/v1/admin/marathons:
    get:
      description: >-
        A diferencia del catálogo, aquí sí salen las no publicadas. Trae el estado declarado
        (`intent`) junto al resuelto (`resolved`), para que se entienda por qué una maratón
        "abierta" aparece llena.
      operationId: AdminController_listarMaratones
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Maratones, publicadas y borradores
      tags:
        - admin
    post:
      description: >-
        Nace **como borrador** salvo que se mande `published: true`: una carrera recién cargada
        suele tener la fecha provisional, y publicarla sola la metería en el catálogo antes de que
        nadie la revise. Sin `slug` se deriva del nombre y se desambigua con un sufijo. Con
        `routeId` la carrera **copia** el recorrido elegido (`GET /admin/routes`): trazado,
        distancia medida y punto de largada. En ese caso `distanceMeters` sobra.
      operationId: AdminController_crearMaraton
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarathonDto'
      responses:
        '201':
          description: La maratón creada, con sus categorías y extras
      security:
        - access-token: []
      summary: Crear una maratón
      tags:
        - admin
...
  /api/v1/admin/marathons/{id}/categories:
    post:
      operationId: AdminController_crearCategoria
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      summary: Agregar una categoría a una maratón
      tags:
        - admin
...
  /api/v1/admin/marathons/{id}/publish:
    post:
      operationId: AdminController_publicar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Publicar una maratón en el catálogo
      tags:
        - admin
...
info:
  title: PaceUp API
  description: |-
    Backend de PaceUp: maratones, inscripciones, planes de entrenamiento y tracking GPS.

    **Unidades base:** distancias en metros, duraciones en segundos, dinero en centavos
    (entero) con `currency`, fechas en ISO-8601 UTC. El formateo (km, min/km, `Bs`) es del cliente.

    **Sobre de respuesta:** todo exito devuelve `{ data, meta }` y todo error `{ error, meta }`.
    Mapea los errores por `error.code`, nunca por `error.message`.
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.paceup.example
    description: VPS
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
    ingest-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: El `ingestToken` que devuelve POST /workouts/sessions
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDto'
        meta:
          $ref: '#/components/schemas/ResponseMetaDto'
      required:
        - error
        - meta
...
    CreateRegistrationDto:
      type: object
      properties:
        marathonId:
          type: string
          description: Id o slug de la maraton
        personalData:
          $ref: '#/components/schemas/PersonalDataDto'
      required:
        - marathonId
        - personalData
...